requirements.txt
dj-database-url==0.4.1
Django==1.8.6
django-bootstrap3==7.0.1
django-bootstrap3-datetimepicker==2.2.3
django-jquery==1.12.2
psycopg2==2.6.1
requests==2.10.0
python_dateutil==2.5.3
geolocation-python==0.2.2